agents/knowledge/requirements.txt
a2a-sdk[http-server]>=0.2.0
google-generativeai>=0.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
agno>=1.0.0
lancedb>=0.15.0
pyarrow>=18.0.0
uvicorn>=0.32.0
fastapi>=0.115.0
